refactor(zratio): move zratio_gauge.h into models/ggm (audit #42)
The gauge header sat at models/ (the model-agnostic tier alongside
base_model.h) yet included models/ggm/zratio_engine.h, the repo's only
layering inversion. The engine and gauge are shared by ggm_model and
mixed_mrf_model equally; place both in models/ggm (mixed reuses the GGM
construct on its continuous subgraph). Include-path change only.
